Make JToolbarHelper::custom calls consistent
# Executive summary
In some part of the code the JToolbarHelper::custom method was called with 6 parameters but the method has only 5 parameters.

# Backwards compatibility
Full b/c, no problems are expected

# Translation impact
none

# Testing instructions
Because the parameter was ignored before there isn’t so much to test. Just make sure the following backend views are still working and the buttons are doing what they did before

I have changed the custom Buttons in this views:

* Extensions/Manage/Update ->Update, Find Updates, Clear Cache
* Extensions/Manage/Discover -> Install, Discover
* Extensions/Manage/Database -> Fix
* Extensions/Manage/Install Languages -> Install, Find Languages
* Components/Joomla! Update -> Clear Cache
* Extensions/Templates -> open a Template:
** Select an Image -> Crop
** Select a Zip Archive (maybe copy a zip into the template Dir) -> Extract Here
** Select a less file -> Compile Less

Thanks to @brianteeman for finding the inconsistency