Allow NULL values for redirect_links new_url column.
New schema update sql scripts for mysql and sqlazure to remove the "NOT
NULL" constraint accidently added with PR #9269 for the new_url column
of table redirect_links in the schema update sql script, so it is
consistent with a new installation. For postgresql it seems there was no
change on the "NOT NULL" constraints because they were not changed with
the column changes.