Let db schema manager use installer script
The schema manager uses script.php anyway just a few lines above calling
the local conversion function, so it can use also the conversion
function of script.php instead of a local one. Both functions are equal
anyway except of the special message text used in script.php, and this
also can be used for the schema manager, the message text might have to
be changed maybe for that.